requirements.txt
# Core dependencies
openai-whisper>=20231117
streamlit>=1.28.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
ffmpeg-python>=0.2.0

# System utilities
pathlib
glob2
psutil

# Optional: For better performance
watchdog>=3.0.0

# Development (optional)
# jupyter
# ipywidgets
